Solution/requirements.txt
apache-airflow==1.10.12
boto3==1.21.32
pandas==1.3.1
psycopg2==2.9.3
